backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
Pillow==10.4.0
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.2
python-dotenv==1.0.1
aiofiles==24.1.0
